\name{prop.comb.RR-package}
\alias{prop.comb.RR-package}
\alias{prop.comb.RR}
\docType{package}
\title{Inferences about a linear combination of proportions}
\description{
prop.comb.RR enables to solve asymptotic inferences (p-values of test and confidence intervals) about any linear combination of independent proportions and the relative risk by means of optimal procedures and the well-known score procedure. It also provides summary measures and a recommendation about the best method that the researcher should choose in each particular case.
}

% \details{
% \tabular{ll}{
% Package: \tab prop.comb.RR\cr
% Type: \tab Package\cr
% Version: \tab 1.0\cr
% Date: \tab 2015-05-19\cr
% License: \tab What license is it under?\cr
% }
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
% }

\author{
Maria Alvarez Hernandez and Javier Roca Pardinas \email{roca@uvigo.es}

Maintainer: Maria Alvarez Hernandez \email{maria.alvarez@uvigo.es}
}

% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
% \keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% # poner ejemplos
%}
